Implemented 2-way refs for widgets and got sorting working

+++ assignment/model/widget/widget.model.server.js
@@ -2,6 +2,7 @@ module.exports = function() {
     var mongoose = require('mongoose');
     var WidgetSchema = require('./widget.schema.server.js')();
     var Widget = mongoose.model('Widget', WidgetSchema);
+    var Page = mongoose.model('Page');
 
     var api = {
         'createWidget': createWidget,
@@ -14,32 +15,62 @@ module.exports = function() {
     return api;
 
     function createWidget(pageId, widget) {
+        widget._page = pageId;
+        widget.rows = widget.rows || 0;
+        widget.size = widget.size || 0;
+        var widgetId = null;
+        var dbWidget = null;
         return Widget
-            .count()
+            .create(widget)
             .then(
-                function(count) {
-                    widget._index = count;
+                function (widget) {
+                    widgetId = widget._id;
+                    dbWidget = widget;
+                    return Page.findById(pageId);
                 },
-                function(error) {
+                function (error) {
                     throw error;
                 })
             .then(
-                function(success) {
-                    widget._page = pageId;
-                    // hacks to prevent undefined cast
-                    widget.rows = widget.rows || 0;
-                    widget.size = widget.size || 0;
-                    return Widget.create(widget);
+                function (page) {
+                    var widgets = page.widgets;
+                    widgets.push(widgetId);
+                    return Page
+                            .update(
+                                { _id: pageId },
+                                { $set:
+                                    {
+                                        widgets: widgets,
+                                    }
+                                });
                 },
-                function(error) {
+                function (error) {
+                    throw error;
+                })
+            .then(
+                function (result) {
+                    if (!result.nModified) {
+                        throw 404;
+                    } else {
+                        return dbWidget;
+                    }
+                },
+                function (error) {
                     throw error;
                 });
     }
 
     function findAllWidgetsForPage(pageId) {
-        return Widget
-            .find({ _page: pageId })
-            .sort({ _index: 'ascending' });
+        return Page
+                .findById(pageId)
+                .populate('widgets')
+                .then(
+                    function (page) {
+                        return page.widgets;
+                    },
+                    function (error) {
+                        throw error;
+                    });
     }
 
     function findWidgetById(widgetId) {
@@ -52,7 +83,7 @@ module.exports = function() {
                     { _id: widgetId },
                     { $set:
                         {
-                            _index: widget._index,
+                            pageIndex: widget.pageIndex,
                             name: widget.name,
                             text: widget.text,
                             placeholder: widget.placeholder,
@@ -71,23 +102,71 @@ module.exports = function() {
     }
 
     function deleteWidget(widgetId) {
-        return Widget.findByIdAndRemove(widgetId);
+        return Widget
+            .findById(widgetId)
+            .then(
+                function (widget) {
+                    return Page.findById(page._widget);
+                },
+                function (error) {
+                    throw error;
+                })
+            .then(
+                function (page) {
+                    var widgets = page.widgets;
+                    var index = widgets.findIndex((e, _i, _a) => e == widgetId);
+                    if (index === -1) {
+                        throw 404;
+                    }
+                    widget.splice(index, 1);
+                    return Page
+                            .update(
+                                { _id: page._id },
+                                { $set:
+                                    {
+                                        widgets: widgets,
+                                    }
+                                });
+                },
+                function (error) {
+                    throw error;
+                })
+            .then(
+                function (result) {
+                    if (!result.nModified) {
+                        throw 404;
+                    }
+                    return Widget.findByIdAndRemove(widgetId);
+                },
+                function (error) {
+                    throw error;
+                });
     }
 
     function reorderWidget(pageId, start, end) {
-        throw 500;
-        // return Widget
-        //     .find({ _page: pageId })
-        //     .and({ $or:[
-        //         { _index: start },
-        //         { _index: end },
-        //     ]})
-        //     .then(
-        //         function(widgets) {
-        //
-        //         },
-        //         function(error) {
-        //
-        //         })
+        return Page
+            .findById(pageId)
+            .then(
+                function (page) {
+                    if (start < 0 ||
+                        end < 0 ||
+                        start >= page.widgets.length ||
+                        end >= page.widgets.length) {
+                            console.log('Widget swap index out of bounds.');
+                            throw 400;
+                        }
+                    widgets = page.widgets;
+                    var tmp = widgets.splice(start, 1)[0];
+                    widgets.splice(end, 0, tmp);
+                    return Page
+                        .update(
+                            { _id: pageId },
+                            { $set: {
+                                widgets: widgets,
+                            }});
+                },
+                function (error) {
+                    throw error;
+                });
     }
 };
